+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 224:
 
 * Remove extraneous error check in advanced-server-flex
 * Advanced servers use HTTP parser interfaces for reading
+* Reusing an HTTP parser returns an error
 
 --------------------------------------------------------------------------------